ARIES-825: Update subsystems to latest Subsystem, Resolver, and Repository APIs.
After much thought, I decided to make the Felix resolver an internal implementation detail of Aries Subsystems for now.

(1) The Felix resolver is the only known implementation of RFC 112 in any shape or form.
(2) The Felix resolver is a work in progress and doesn't reflect the very latest API changes, which are used in the CT. Consequently, it is not possible to run against the CT and import the API without updating the Felix resolver. The task of updating the Felix resolver code to the latest API is not trivial and would be throw away work as Felix is in the process of updating it themselves.
(3) The Felix resolver is in an SCM sandbox and not currently published to a maven repository.

Once Felix has updated their implementation and made it available in a maven repository, I will back this out and switch to it.

git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1245173 13f79535-47bb-0310-9956-ffa450edef68