ARIES-825: Update subsystems to latest Subsystem, Resolver, and Repository APIs.
(1) Added initial support for composite Import-Package header.
(2) Added new composite test.
(3) Updated references to coordinator and region in subsystem pom.
(4) Removed the installation of a temporary bundle in order to avoid location renaming as part of installing the region context bundle as it is no longer necessary with the latest region updates.
(5) Switched back to org.osgi.framework.VersionRange since there's a public build of equinox with it now.
(6) Fixed some more issues with the regex grammar not accounting for whitespace between terminals.

git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1293711 13f79535-47bb-0310-9956-ffa450edef68