The equinox38 profile will not execute on the build machine even when active by default. This might be because the default activation policy is being overridden by other profiles being specified somewhere else. I don't see this occurring in the build configuration (e.g. via the command line), but there are other places it could be happening. The profile activates fine locally, which I found confusing. However, I then noticed that I'm using maven 3.0.4 locally whereas the build uses maven 2.2.1. I need to test with 2.2.1 locally to see if it can be reproduced.
In the meantime, I removed the profiles and put the maven-paxexam-plugin in the general <build><plugins> section where it will always be executed. To switch from one equinox version to another, the desired version would need to be uncommented in the POM.

Another option would be to reinsert the profiles and specify the profile or a system property as part of the build configuration.

git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1396352 13f79535-47bb-0310-9956-ffa450edef68