Updates for preexisting regions.
The root region constant has been removed since the root region is now determined by the region in which the subsystem.core bundle
was installed.

Root subsystem initialization and capability validation now take preexisting regions into account.

New test to insure resources within preexisting regions can be capability providers if the sharing policy allows.

The specification takes the point of view that subsystems is the only isolation engine existing in the framework. These changes
provide leniency for the Virgo use case where there are two preexisting regions, kernel and user, and subsystems is installed into
the user region. Although the changes were made in the spirit of generality, no attempt was made to introduce a compatibility
mechanism for multiple isolation engines in the same framework. However, it might make sense to direct some future work towards
providing a more robust compability with extraneous Region Digraph regions, since this is the isolation engine used by the
implementation. For example, a "regions as subsystems" approach could be taken where a subsystem is created for each preexisting
region. This could be problematic and potentially complex, however, since the subsystem graph can only handle a subset of
graphs that are possible when using Digraph directly.

One other thing to note about these changes is that they implement the spec wording literally, which states that any bundles
installed out of band (i.e. outside of the subsystems API) become constituents of the root subsystem. This means that all
bundles that are part of preexisting regions will become constituents of the root subsystem, even if they are not actually in
the root subsystem region. This results in a potential disparity between the bundles returned by getConstituents() and
getBundleContext().getBundles(). We may wish to address this disparity at both the specification and implementation level in
the future.

git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1401402 13f79535-47bb-0310-9956-ffa450edef68