ARIES-1429 NullPointerException at org.apache.aries.subsystem.core.internal.WovenClassListener.modified at org.apache.aries.subsystem.core.internal.RegionUpdater.addRequirements
Add tests for both the woven class listener and AriesSubsystem.addRequirements.

A missing connection between parent and child may be tolerated for the woven class listener but not when adding requirements to the sharing policy. The region updater will now
throw an IllegalStateException when it detects a missing edge between the head and tail (parent and child) regions. The woven class listener will catch and ignore this
exception. When adding requirements, however, this exception will be propogated as before.

git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1707803 13f79535-47bb-0310-9956-ffa450edef68