[ARIES-1503] Smarter fix for cxf
CXF jaxws namespace handler imports the standard cxf namespaces, but its namespace handler always return a url to the jaxws xsd, which is not used at all.
In order to work around those cases, we do allow the dynamic addition of namespaces while building the schema.  If a schema can’t be found, we raise a MissingNamespaceException which is handled by the BlueprintContainerImpl and keep the container in the WaitForNamespaceHandlers state.

git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1738872 13f79535-47bb-0310-9956-ffa450edef68